created separate file for fee integration test
