+++ bridge-token-factory/tests/token_transfer.rs
@@ -2,7 +2,6 @@ use bridge_token_factory::{validate_eth_address, EthLockedEvent, Proof};
 use near_sdk::borsh::{self, BorshSerialize};
 use near_sdk::{AccountId, Balance, ONE_NEAR, ONE_YOCTO};
 use serde_json::json;
-use tokio::runtime::Runtime;
 use workspaces::result::ExecutionFinalResult;
 use workspaces::{network::Sandbox, Account, Contract, Worker};
 
@@ -13,77 +12,52 @@ const SENDER_ADDRESS: &str = "00005474e89094c44da98b954eedeac495271d0f";
 const ALICE: &str = "alice.test.near";
 
 const FACTORY_WASM_PATH: &str = "../res/bridge_token_factory.wasm";
+const FACTORY_WASM_PATH_V_0_1_6: &str = "res/bridge_token_factory_v0.1.6.wasm";
+const BRIDGE_TOKEN_WASM_PATH: &str = "../res/bridge_token.wasm";
 const MOCK_PROVER_WASM_PATH: &str = "../res/mock_prover.wasm";
 
 const DEFAULT_DEPOSIT: u128 = ONE_NEAR;
 
-fn create_contract() -> (Account, Contract, Account, Worker<Sandbox>) {
-    let rt = Runtime::new().unwrap();
-
-    let worker: Worker<Sandbox> = rt.block_on(workspaces::sandbox()).unwrap();
+async fn create_contract(factory_wasm_path: &str) -> (Account, Contract, Worker<Sandbox>) {
+    let worker: Worker<Sandbox> = workspaces::sandbox().await.unwrap();
     let prover_wasm = std::fs::read(MOCK_PROVER_WASM_PATH).unwrap();
-    let prover_contract: Contract = rt.block_on(worker.dev_deploy(&prover_wasm)).unwrap();
+    let prover_contract: Contract = worker.dev_deploy(&prover_wasm).await.unwrap();
     let prover_id = prover_contract.id();
 
     let owner = worker.root_account().unwrap();
-    let factory_account = rt
-        .block_on(
-            owner
-                .create_subaccount(FACTORY)
-                .initial_balance(200 * ONE_NEAR)
-                .transact(),
-        )
+    let factory_account = owner
+        .create_subaccount(FACTORY)
+        .initial_balance(200 * ONE_NEAR)
+        .transact()
+        .await
         .unwrap()
         .into_result()
         .unwrap();
 
-    let factory_wasm = std::fs::read(FACTORY_WASM_PATH).unwrap();
-    let factory_contract: Contract = rt
-        .block_on(factory_account.deploy(&factory_wasm))
-        .unwrap()
-        .result;
+    let factory_wasm = std::fs::read(factory_wasm_path).unwrap();
+    let factory_contract: Contract = factory_account.deploy(&factory_wasm).await.unwrap().result;
 
-    let alice = rt
-        .block_on(
-            owner
-                .create_subaccount("alice")
-                .initial_balance(200 * ONE_NEAR)
-                .transact(),
-        )
+    let alice = owner
+        .create_subaccount("alice")
+        .initial_balance(200 * ONE_NEAR)
+        .transact()
+        .await
         .unwrap()
         .into_result()
         .unwrap();
 
-    let _result = rt.block_on(
-        factory_account
-            .call(factory_contract.id(), "new")
-            .args(
-                json!({"prover_account": prover_id, "locker_address": LOCKER_ADDRESS.to_string()})
-                    .to_string()
-                    .into_bytes(),
-            )
-            .transact(),
-    )
-    .unwrap();
-
-    let _grant_fee_setter_call = rt
-        .block_on(
-            factory_account
-                .call(factory_contract.id(), "acl_grant_role")
-                .args(
-                    json!({"role": "FeeSetter".to_string(), "account_id": alice.id()})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .transact(),
+    let _result = factory_account
+        .call(factory_contract.id(), "new")
+        .args(
+            json!({"prover_account": prover_id, "locker_address": LOCKER_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes(),
         )
+        .transact()
+        .await
         .unwrap();
-    assert!(
-        _grant_fee_setter_call.is_success(),
-        "fee setter grant role failed"
-    );
 
-    (alice, factory_contract, factory_account, worker)
+    (alice, factory_contract, worker)
 }
 
 #[derive(BorshSerialize)]
@@ -121,76 +95,55 @@ fn remove_quotes(s: &mut String) -> String {
     s.to_string()
 }
 
-fn run_view_function(
-    rt: &Runtime,
+async fn run_view_function(
     contract_id: &String,
     worker: &Worker<Sandbox>,
     function: &str,
     args: serde_json::Value,
 ) -> String {
     let mut res = std::str::from_utf8(
-        &rt.block_on(worker.view(
-            &contract_id.parse().unwrap(),
-            function,
-            args.to_string().into_bytes(),
-        ))
-        .unwrap()
-        .result,
+        &worker
+            .view(
+                &contract_id.parse().unwrap(),
+                function,
+                args.to_string().into_bytes(),
+            )
+            .await
+            .unwrap()
+            .result,
     )
     .unwrap()
     .to_string();
     remove_quotes(&mut res)
 }
 
-fn get_fee_amount(
-    amount: u128,
-    fee_percentage: u128,
-    lower_bound: u128,
-    upper_bound: u128,
-) -> u128 {
-    const FEE_DECIMAL_PRECISION: u128 = 100_00_00;
-    let mut fee_amount = (amount * fee_percentage) / FEE_DECIMAL_PRECISION;
-    if fee_amount < lower_bound {
-        fee_amount = lower_bound;
-        return fee_amount;
-    } else if fee_amount > upper_bound {
-        fee_amount = upper_bound;
-        return fee_amount;
-    } else {
-        return fee_amount;
-    }
-}
-
-#[test]
-fn test_token_transfer_with_deposit_and_withdraw_fee() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
+#[tokio::test]
+async fn test_eth_token_transfer() {
+    let (alice, factory, worker) = create_contract(FACTORY_WASM_PATH).await;
     const INIT_ALICE_BALANCE: u64 = 1000;
     const WITHDRAW_AMOUNT: u64 = 100;
 
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
+    assert!(alice
+        .call(factory.id(), "deploy_bridge_token")
+        .deposit(35 * ONE_NEAR)
+        .args(
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes()
         )
+        .max_gas()
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     let token_account_id: String = run_view_function(
-        &rt,
         &factory.id().to_string(),
         &worker,
         "get_bridge_token_account_id",
         json!({"address": DAI_ADDRESS.to_string()}),
-    );
+    )
+    .await;
 
     assert_eq!(
         token_account_id,
@@ -198,22 +151,17 @@ fn test_token_transfer_with_deposit_and_withdraw_fee() {
     );
 
     let alice_balance: String = run_view_function(
-        &rt,
         &token_account_id,
         &worker,
         "ft_balance_of",
         json!({ "account_id": ALICE }),
-    );
+    )
+    .await;
 
     assert_eq!(alice_balance, "0");
 
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
+    let total_supply: String =
+        run_view_function(&token_account_id, &worker, "ft_total_supply", json!({})).await;
     assert_eq!(total_supply, "0");
 
     let mut proof = Proof::default();
@@ -226,1624 +174,91 @@ fn test_token_transfer_with_deposit_and_withdraw_fee() {
     }
     .to_log_entry_data();
 
-    let fee_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_bound")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 500, "lower_bound": 100})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-    assert!(fee_setter_call.is_success(), "Fee setter called failed");
-
-    let fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "eth_to_near": 400000, "eth_to_aurora": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        fee_percentage_setter_call.is_success(),
-        "Fee percentage setter called failed"
-    );
-
-    let withdraw_fee_setter_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "set_withdraw_fee_bound")
-                .args(
-                    json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 50, "lower_bound": 10})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .max_gas()
-                .transact(),
-        )
-        .unwrap();
-    assert!(
-        withdraw_fee_setter_call.is_success(),
-        "Withdraw Fee setter called failed"
-    );
-
-    let withdraw_fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_withdraw_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "near_to_eth": 350000, "aurora_to_eth": 300000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        withdraw_fee_percentage_setter_call.is_success(),
-        "Withdraw fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-    let deposit_fee_amount =
-        get_fee_amount(u128::from(INIT_ALICE_BALANCE), 400000u128, 100u128, 500u128);
-    let transfer_amount = INIT_ALICE_BALANCE as u128 - deposit_fee_amount;
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", transfer_amount));
-
-    let token_factory_balance_after_deposit: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
-    );
-    assert_eq!(
-        token_factory_balance_after_deposit,
-        format!("{}", deposit_fee_amount)
-    );
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : format!("{}", WITHDRAW_AMOUNT),
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        )
+    assert!(alice
+        .call(factory.id(), "deposit")
+        .deposit(DEFAULT_DEPOSIT)
+        .max_gas()
+        .args(proof.try_to_vec().unwrap())
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     let alice_balance: String = run_view_function(
-        &rt,
         &token_account_id,
         &worker,
         "ft_balance_of",
         json!({ "account_id": ALICE }),
-    );
-    assert_eq!(
-        alice_balance,
-        format!(
-            "{}",
-            u128::from(INIT_ALICE_BALANCE) - u128::from(WITHDRAW_AMOUNT) - deposit_fee_amount
-        )
-    );
-    let withdraw_fee_amount =
-        get_fee_amount(u128::from(WITHDRAW_AMOUNT), 350000u128, 10u128, 50u128);
-    let token_factory_balance_after_withdraw: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string()}),
-    );
-    assert_eq!(
-        format!("{}", deposit_fee_amount + withdraw_fee_amount),
-        token_factory_balance_after_withdraw
-    );
-}
-
-#[test]
-fn test_token_deposit_without_fee_bound_and_fee_percentage() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
+    )
+    .await;
     assert_eq!(alice_balance, format!("{}", INIT_ALICE_BALANCE));
 
-    let token_factory_balance_after_deposit: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string()}),
-    );
-    assert_eq!(token_factory_balance_after_deposit, format!("{}", 0));
-}
+    let total_supply: String =
+        run_view_function(&token_account_id, &worker, "ft_total_supply", json!({})).await;
+    assert_eq!(total_supply, format!("{}", INIT_ALICE_BALANCE));
 
-#[test]
-fn test_token_deposit_with_fee_less_than_lower_bound() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
+    assert!(alice
+        .call(&token_account_id.parse().unwrap(), "withdraw")
+        .max_gas()
+        .deposit(ONE_YOCTO)
+        .args(
+            json!({
+                "amount" : format!("{}", WITHDRAW_AMOUNT),
+                "recipient" : SENDER_ADDRESS.to_string()
+            })
+            .to_string()
+            .into_bytes(),
         )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let fee_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_bound")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 200, "lower_bound": 100})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-    assert!(fee_setter_call.is_success(), "Fee setter called failed");
-
-    let fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "eth_to_near": 50000, "eth_to_aurora": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        fee_percentage_setter_call.is_success(),
-        "Fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-    let fee_amount = get_fee_amount(u128::from(INIT_ALICE_BALANCE), 50000u128, 100u128, 200u128);
-    let transfer_amount = INIT_ALICE_BALANCE as u128 - fee_amount;
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", transfer_amount));
-
-    let token_factory_balance_after_deposit: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string()}),
-    );
-    assert_eq!(
-        token_factory_balance_after_deposit,
-        format!("{}", fee_amount)
-    );
-}
-
-#[test]
-fn test_token_deposit_with_fee_more_than_upper_bound() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let fee_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_bound")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 200, "lower_bound": 100})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-    assert!(fee_setter_call.is_success(), "Fee setter called failed");
-
-    let fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "eth_to_near": 400000, "eth_to_aurora": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        fee_percentage_setter_call.is_success(),
-        "Fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-    let fee_amount = get_fee_amount(u128::from(INIT_ALICE_BALANCE), 400000u128, 100u128, 200u128);
-    let transfer_amount = INIT_ALICE_BALANCE as u128 - fee_amount;
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", transfer_amount));
-
-    let token_factory_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string()}),
-    );
-    assert_eq!(token_factory_balance, format!("{}", fee_amount));
-}
-
-#[test]
-fn test_token_deposit_with_fee_in_bound_range() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let fee_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_bound")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 500, "lower_bound": 100})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-    assert!(fee_setter_call.is_success(), "Fee setter called failed");
-
-    let fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "eth_to_near": 400000, "eth_to_aurora": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        fee_percentage_setter_call.is_success(),
-        "Fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-    let fee_amount = get_fee_amount(u128::from(INIT_ALICE_BALANCE), 400000u128, 100u128, 500u128);
-    let transfer_amount = INIT_ALICE_BALANCE as u128 - fee_amount;
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", transfer_amount));
-
-    let token_factory_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
-    );
-    assert_eq!(token_factory_balance, format!("{}", fee_amount));
-}
-
-#[test]
-fn test_token_withdraw_without_fee_bound_and_fee_percentage() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-    const WITHDRAW_AMOUNT: u64 = 100;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", INIT_ALICE_BALANCE));
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : format!("{}", WITHDRAW_AMOUNT),
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        )
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     let alice_balance: String = run_view_function(
-        &rt,
         &token_account_id,
         &worker,
         "ft_balance_of",
         json!({ "account_id": ALICE }),
-    );
+    )
+    .await;
     assert_eq!(
         alice_balance,
-        format!(
-            "{}",
-            u128::from(INIT_ALICE_BALANCE) - u128::from(WITHDRAW_AMOUNT)
-        )
+        format!("{}", INIT_ALICE_BALANCE - WITHDRAW_AMOUNT)
     );
-    let token_factory_balance_after_withdraw: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
+
+    let total_supply: String =
+        run_view_function(&token_account_id, &worker, "ft_total_supply", json!({})).await;
+    assert_eq!(
+        total_supply,
+        format!("{}", INIT_ALICE_BALANCE - WITHDRAW_AMOUNT)
     );
-    assert_eq!(format!("{}", 0), token_factory_balance_after_withdraw);
 }
 
-#[test]
-fn test_token_withdraw_with_fee_less_than_lower_bound() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-    const WITHDRAW_AMOUNT: u64 = 100;
+#[tokio::test]
+async fn test_with_invalid_proof() {
+    let (user, factory, worker) = create_contract(FACTORY_WASM_PATH).await;
 
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
+    assert!(user
+        .call(factory.id(), "deploy_bridge_token")
+        .deposit(35 * ONE_NEAR)
+        .args(
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes()
         )
+        .max_gas()
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     let token_account_id: String = run_view_function(
-        &rt,
         &factory.id().to_string(),
         &worker,
         "get_bridge_token_account_id",
         json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let withdraw_fee_setter_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "set_withdraw_fee_bound")
-                .args(
-                    json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 20, "lower_bound": 10})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .max_gas()
-                .transact(),
-        )
-        .unwrap();
-    assert!(
-        withdraw_fee_setter_call.is_success(),
-        "Withdraw Fee setter called failed"
-    );
-
-    let withdraw_fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_withdraw_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "near_to_eth": 50000, "aurora_to_eth": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
     )
-    .unwrap();
-
-    assert!(
-        withdraw_fee_percentage_setter_call.is_success(),
-        "Withdraw fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", INIT_ALICE_BALANCE));
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : format!("{}", WITHDRAW_AMOUNT),
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        )
-        .unwrap()
-        .is_success());
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(
-        alice_balance,
-        format!(
-            "{}",
-            u128::from(INIT_ALICE_BALANCE) - u128::from(WITHDRAW_AMOUNT)
-        )
-    );
-    let fee_amount = get_fee_amount(u128::from(WITHDRAW_AMOUNT), 50000u128, 10u128, 50u128);
-    let token_factory_balance_after_withdraw: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
-    );
-    assert_eq!(
-        format!("{}", fee_amount),
-        token_factory_balance_after_withdraw
-    );
-}
-#[test]
-fn test_token_withdraw_with_fee_more_than_upper_bound() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-    const WITHDRAW_AMOUNT: u64 = 100;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let withdraw_fee_setter_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "set_withdraw_fee_bound")
-                .args(
-                    json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 30, "lower_bound": 10})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .max_gas()
-                .transact(),
-        )
-        .unwrap();
-    assert!(
-        withdraw_fee_setter_call.is_success(),
-        "Withdraw Fee setter called failed"
-    );
-
-    let withdraw_fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_withdraw_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "near_to_eth": 350000, "aurora_to_eth": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        withdraw_fee_percentage_setter_call.is_success(),
-        "Withdraw fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", INIT_ALICE_BALANCE));
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : format!("{}", WITHDRAW_AMOUNT),
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        )
-        .unwrap()
-        .is_success());
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(
-        alice_balance,
-        format!(
-            "{}",
-            u128::from(INIT_ALICE_BALANCE) - u128::from(WITHDRAW_AMOUNT)
-        )
-    );
-    let fee_amount = get_fee_amount(u128::from(WITHDRAW_AMOUNT), 350000u128, 10u128, 30u128);
-    let token_factory_balance_after_withdraw: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
-    );
-    assert_eq!(
-        format!("{}", fee_amount),
-        token_factory_balance_after_withdraw
-    );
-}
-
-#[test]
-fn test_token_withdraw_with_fee_in_bound_range() {
-    let (alice, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-    const WITHDRAW_AMOUNT: u64 = 100;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let withdraw_fee_setter_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "set_withdraw_fee_bound")
-                .args(
-                    json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 50, "lower_bound": 10})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .max_gas()
-                .transact(),
-        )
-        .unwrap();
-    assert!(
-        withdraw_fee_setter_call.is_success(),
-        "Withdraw Fee setter called failed"
-    );
-
-    let withdraw_fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_withdraw_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "near_to_eth": 400000, "aurora_to_eth": 400000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        withdraw_fee_percentage_setter_call.is_success(),
-        "Withdraw fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", INIT_ALICE_BALANCE));
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : format!("{}", WITHDRAW_AMOUNT),
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        )
-        .unwrap()
-        .is_success());
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(
-        alice_balance,
-        format!(
-            "{}",
-            u128::from(INIT_ALICE_BALANCE) - u128::from(WITHDRAW_AMOUNT)
-        )
-    );
-    let fee_amount = get_fee_amount(u128::from(WITHDRAW_AMOUNT), 400000u128, 10u128, 50u128);
-    let token_factory_balance_after_withdraw: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
-    );
-    assert_eq!(
-        format!("{}", fee_amount),
-        token_factory_balance_after_withdraw
-    );
-}
-
-#[test]
-fn test_fee_deposit_claim() {
-    let (alice, factory, factory_account, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-    const INIT_ALICE_BALANCE: u64 = 1000;
-    const WITHDRAW_AMOUNT: u64 = 100;
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
-
-    assert_eq!(
-        token_account_id,
-        format!("{}.{}", DAI_ADDRESS, factory.id())
-    );
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    assert_eq!(alice_balance, "0");
-
-    let total_supply: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_total_supply",
-        json!({}),
-    );
-    assert_eq!(total_supply, "0");
-
-    let _grant_fee_claimer_role_call = rt
-        .block_on(
-            factory_account
-                .call(factory.id(), "acl_grant_role")
-                .args(
-                    json!({"role": "FeeClaimer".to_string(), "account_id": alice.id()})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .transact(),
-        )
-        .unwrap();
-    assert!(
-        _grant_fee_claimer_role_call.is_success(),
-        "fee setter grant role failed"
-    );
-
-    let mut proof = Proof::default();
-    proof.log_entry_data = EthLockedEvent {
-        locker_address: validate_eth_address(LOCKER_ADDRESS.to_string()),
-        token: DAI_ADDRESS.to_string(),
-        sender: SENDER_ADDRESS.to_string(),
-        amount: Balance::from(INIT_ALICE_BALANCE),
-        recipient: ALICE.parse().unwrap(),
-    }
-    .to_log_entry_data();
-
-    let fee_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_bound")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 500, "lower_bound": 100})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-    assert!(fee_setter_call.is_success(), "Fee setter called failed");
-
-    let fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_deposit_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "eth_to_near": 400000, "eth_to_aurora": 40000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        fee_percentage_setter_call.is_success(),
-        "Fee percentage setter called failed"
-    );
-
-    let withdraw_fee_setter_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "set_withdraw_fee_bound")
-                .args(
-                    json!({"token": DAI_ADDRESS.to_string(), "upper_bound": 50, "lower_bound": 10})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .max_gas()
-                .transact(),
-        )
-        .unwrap();
-    assert!(
-        withdraw_fee_setter_call.is_success(),
-        "Withdraw Fee setter called failed"
-    );
-
-    let withdraw_fee_percentage_setter_call = &rt
-    .block_on(
-        alice
-            .call(factory.id(), "set_withdraw_fee_percentage")
-            .args(
-                json!({"token": DAI_ADDRESS.to_string(), "near_to_eth": 350000, "aurora_to_eth": 300000})
-                    .to_string()
-                    .into_bytes()
-            )
-            .max_gas()
-            .transact()
-    )
-    .unwrap();
-
-    assert!(
-        withdraw_fee_percentage_setter_call.is_success(),
-        "Withdraw fee percentage setter called failed"
-    );
-
-    let deposit_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "deposit")
-                .deposit(DEFAULT_DEPOSIT)
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        )
-        .unwrap();
-    if deposit_call.is_failure() {
-        println!("\n\n\n Deposit error {:?}\n\n", deposit_call.failures());
-    }
-    let deposit_fee_amount =
-        get_fee_amount(u128::from(INIT_ALICE_BALANCE), 400000u128, 100u128, 500u128);
-    let transfer_amount = INIT_ALICE_BALANCE as u128 - deposit_fee_amount;
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(alice_balance, format!("{}", transfer_amount));
-
-    let token_factory_balance_after_deposit: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string() }),
-    );
-    assert_eq!(
-        token_factory_balance_after_deposit,
-        format!("{}", deposit_fee_amount)
-    );
-
-    assert!(&rt
-        .block_on(
-            alice
-                .call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : format!("{}", WITHDRAW_AMOUNT),
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        )
-        .unwrap()
-        .is_success());
-
-    let alice_balance: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    assert_eq!(
-        alice_balance,
-        format!(
-            "{}",
-            u128::from(INIT_ALICE_BALANCE) - u128::from(WITHDRAW_AMOUNT) - deposit_fee_amount
-        )
-    );
-    let withdraw_fee_amount =
-        get_fee_amount(u128::from(WITHDRAW_AMOUNT), 350000u128, 10u128, 50u128);
-    let token_factory_balance_after_withdraw: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string()}),
-    );
-    assert_eq!(
-        format!("{}", deposit_fee_amount + withdraw_fee_amount),
-        token_factory_balance_after_withdraw
-    );
-
-    let alice_balance_before_claim: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-
-    let claim_fee_call = &rt
-        .block_on(
-            alice
-                .call(factory.id(), "claim_fee")
-                .args(
-                    json!({"token": token_account_id, "amount": 50})
-                        .to_string()
-                        .into_bytes(),
-                )
-                .max_gas()
-                .transact(),
-        )
-        .unwrap();
-    assert!(claim_fee_call.is_success(), "Claim Fee call failed");
-
-    let token_factory_balance_after_claim: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": factory.id().to_string()}),
-    );
-
-    let alice_balance_after_claim: String = run_view_function(
-        &rt,
-        &token_account_id,
-        &worker,
-        "ft_balance_of",
-        json!({ "account_id": ALICE }),
-    );
-    let alice_bal_before_claim: u128 = alice_balance_before_claim.parse().unwrap();
-    let token_factory_balance_before_claim: u128 =
-        token_factory_balance_after_withdraw.parse().unwrap();
-    assert_eq!(
-        format!("{}", token_factory_balance_before_claim - 50),
-        token_factory_balance_after_claim,
-        "Token Factory balance didn't matched before and after fee is claimed"
-    );
-    assert_eq!(
-        format!("{}", alice_bal_before_claim + 50),
-        alice_balance_after_claim,
-        "Before-After claim balance of alice not matched"
-    );
-}
-
-#[test]
-fn test_with_invalid_proof() {
-    let (user, factory, _, worker) = create_contract();
-    let rt = Runtime::new().unwrap();
-
-    assert!(&rt
-        .block_on(
-            user.call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
-        )
-        .unwrap()
-        .is_success());
-
-    let token_account_id: String = run_view_function(
-        &rt,
-        &factory.id().to_string(),
-        &worker,
-        "get_bridge_token_account_id",
-        json!({"address": DAI_ADDRESS.to_string()}),
-    );
+    .await;
 
     assert_eq!(
         token_account_id,
@@ -1866,12 +281,12 @@ fn test_with_invalid_proof() {
     proof.proof = vec![vec![]];
 
     assert_error(
-        &rt.block_on(
-            user.call(factory.id(), "deposit")
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        ),
+        &user
+            .call(factory.id(), "deposit")
+            .max_gas()
+            .args(proof.try_to_vec().unwrap())
+            .transact()
+            .await,
         "Failed to verify the proof",
     );
 
@@ -1880,47 +295,44 @@ fn test_with_invalid_proof() {
 
     // This deposit event must succeed. Notice that previously a similar deposit
     // was made, but it failed because it had an invalid proof, so this one should succeed.
-    assert!(&rt
-        .block_on(
-            user.call(factory.id(), "deposit")
-                .max_gas()
-                .deposit(DEFAULT_DEPOSIT)
-                .args(proof.try_to_vec().unwrap())
-                .transact()
-        )
+    assert!(user
+        .call(factory.id(), "deposit")
+        .max_gas()
+        .deposit(DEFAULT_DEPOSIT)
+        .args(proof.try_to_vec().unwrap())
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     // This deposit event must fail since same deposit event can't be reused.
     // Previous call to deposit with the same event was successful.
     assert_error(
-        &rt.block_on(
-            user.call(factory.id(), "deposit")
-                .max_gas()
-                .args(proof.try_to_vec().unwrap())
-                .transact(),
-        ),
+        &user
+            .call(factory.id(), "deposit")
+            .max_gas()
+            .args(proof.try_to_vec().unwrap())
+            .transact()
+            .await,
         "Event cannot be reused for depositing.",
     );
 }
 
-#[test]
-fn test_bridge_token_failures() {
-    let (user, factory, _, _worker) = create_contract();
-    let rt = Runtime::new().unwrap();
+#[tokio::test]
+async fn test_bridge_token_failures() {
+    let (user, factory, _worker) = create_contract(FACTORY_WASM_PATH).await;
 
-    assert!(&rt
-        .block_on(
-            user.call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
+    assert!(user
+        .call(factory.id(), "deploy_bridge_token")
+        .deposit(35 * ONE_NEAR)
+        .args(
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes()
         )
+        .max_gas()
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
@@ -1928,20 +340,20 @@ fn test_bridge_token_failures() {
 
     // Fail to withdraw because the account is not registered (and have no coins)
     assert_error(
-        &rt.block_on(
-            user.call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .deposit(ONE_YOCTO)
-                .args(
-                    json!({
-                        "amount" : "100",
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .transact(),
-        ),
+        &user
+            .call(&token_account_id.parse().unwrap(), "withdraw")
+            .max_gas()
+            .deposit(ONE_YOCTO)
+            .args(
+                json!({
+                    "amount" : "100",
+                    "recipient" : SENDER_ADDRESS.to_string()
+                })
+                .to_string()
+                .into_bytes(),
+            )
+            .transact()
+            .await,
         " is not registered",
     );
 
@@ -1949,127 +361,313 @@ fn test_bridge_token_failures() {
     let account_id: Option<AccountId> = None;
     let registration_only: Option<bool> = None;
 
-    assert!(&rt
-        .block_on(
-            user.call(&token_account_id.parse().unwrap(), "storage_deposit")
-                .max_gas()
-                .args(
-                    json!({
-                        "account_id": account_id,
-                        "registration_only": registration_only
-                    })
-                    .to_string()
-                    .into_bytes()
-                )
-                .deposit(DEFAULT_DEPOSIT)
-                .transact()
+    assert!(user
+        .call(&token_account_id.parse().unwrap(), "storage_deposit")
+        .max_gas()
+        .args(
+            json!({
+                "account_id": account_id,
+                "registration_only": registration_only
+            })
+            .to_string()
+            .into_bytes()
         )
+        .deposit(DEFAULT_DEPOSIT)
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     // Fail to withdraw because the account has no enough balance
     assert_error(
-        &rt.block_on(
-            user.call(&token_account_id.parse().unwrap(), "withdraw")
-                .max_gas()
-                .args(
-                    json!({
-                        "amount" : "100",
-                        "recipient" : SENDER_ADDRESS.to_string()
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .deposit(ONE_YOCTO)
-                .transact(),
-        ),
+        &user
+            .call(&token_account_id.parse().unwrap(), "withdraw")
+            .max_gas()
+            .args(
+                json!({
+                    "amount" : "100",
+                    "recipient" : SENDER_ADDRESS.to_string()
+                })
+                .to_string()
+                .into_bytes(),
+            )
+            .deposit(ONE_YOCTO)
+            .transact()
+            .await,
         "The account doesn't have enough balance",
     );
 
-    let other_user = rt
-        .block_on(
-            user.create_subaccount("bob")
-                .initial_balance(50 * ONE_NEAR)
-                .transact(),
-        )
+    let other_user = user
+        .create_subaccount("bob")
+        .initial_balance(50 * ONE_NEAR)
+        .transact()
+        .await
         .unwrap()
         .into_result()
         .unwrap();
 
     // Fail to mint because the caller is not the controller
     assert_error(
-        &rt.block_on(
-            other_user
-                .call(&token_account_id.parse().unwrap(), "mint")
-                .max_gas()
-                .args(
-                    json!({
-                        "account_id" : ALICE.to_string(),
-                        "amount" : "100"
-                    })
-                    .to_string()
-                    .into_bytes(),
-                )
-                .deposit(ONE_YOCTO)
-                .transact(),
-        ),
+        &other_user
+            .call(&token_account_id.parse().unwrap(), "mint")
+            .max_gas()
+            .args(
+                json!({
+                    "account_id" : ALICE.to_string(),
+                    "amount" : "100"
+                })
+                .to_string()
+                .into_bytes(),
+            )
+            .deposit(ONE_YOCTO)
+            .transact()
+            .await,
         "Only controller can call mint",
     );
 }
 
-#[test]
-fn test_deploy_failures() {
-    let (user, factory, _, _worker) = create_contract();
-    let rt = Runtime::new().unwrap();
+#[tokio::test]
+async fn test_deploy_failures() {
+    let (user, factory, _worker) = create_contract(FACTORY_WASM_PATH).await;
 
     // Fails with not enough deposit.
     assert_error(
-        &rt.block_on(
-            user.call(factory.id(), "deploy_bridge_token")
-                .deposit(0)
-                .args(json!({ "address": DAI_ADDRESS }).to_string().into_bytes())
-                .max_gas()
-                .transact(),
-        ),
+        &user
+            .call(factory.id(), "deploy_bridge_token")
+            .deposit(0)
+            .args(json!({ "address": DAI_ADDRESS }).to_string().into_bytes())
+            .max_gas()
+            .transact()
+            .await,
         "Not enough attached deposit to complete bridge token creation",
     );
 
     // Fails with address is invalid.
     assert_error(
-        &rt.block_on(
-            user.call(factory.id(), "deploy_bridge_token")
-                .deposit(0)
-                .args(json!({"address": "not_a_hex"}).to_string().into_bytes())
-                .max_gas()
-                .transact(),
-        ),
+        &user
+            .call(factory.id(), "deploy_bridge_token")
+            .deposit(0)
+            .args(json!({"address": "not_a_hex"}).to_string().into_bytes())
+            .max_gas()
+            .transact()
+            .await,
         "address should be a valid hex string.: OddLength",
     );
 
     // Fails second time because already exists.
-    assert!(&rt
-        .block_on(
-            user.call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(
-                    json!({"address": DAI_ADDRESS.to_string()})
-                        .to_string()
-                        .into_bytes()
-                )
-                .max_gas()
-                .transact()
+    assert!(user
+        .call(factory.id(), "deploy_bridge_token")
+        .deposit(35 * ONE_NEAR)
+        .args(
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes()
         )
+        .max_gas()
+        .transact()
+        .await
         .unwrap()
         .is_success());
 
     assert_error(
-        &rt.block_on(
-            user.call(factory.id(), "deploy_bridge_token")
-                .deposit(35 * ONE_NEAR)
-                .args(json!({ "address": DAI_ADDRESS }).to_string().into_bytes())
-                .max_gas()
-                .transact(),
-        ),
+        &user
+            .call(factory.id(), "deploy_bridge_token")
+            .deposit(35 * ONE_NEAR)
+            .args(json!({ "address": DAI_ADDRESS }).to_string().into_bytes())
+            .max_gas()
+            .transact()
+            .await,
         "BridgeToken contract already exists.",
     );
+}
+
+#[tokio::test]
+async fn test_upgrade() {
+    let (alice, factory, worker) = create_contract(FACTORY_WASM_PATH_V_0_1_6).await;
+    const INIT_ALICE_BALANCE: u64 = 1000;
+    const WITHDRAW_AMOUNT: u64 = 100;
+
+    let token_account = Account::from_secret_key(
+        format!("{}.{}", DAI_ADDRESS, factory.id()).parse().unwrap(),
+        factory.as_account().secret_key().clone(),
+        &worker,
+    );
+
+    assert!(alice
+        .call(factory.id(), "deploy_bridge_token")
+        .deposit(35 * ONE_NEAR)
+        .args(
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes()
+        )
+        .max_gas()
+        .transact()
+        .await
+        .unwrap()
+        .is_success());
+
+    let token_account_id: String = factory
+        .view(
+            "get_bridge_token_account_id",
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes(),
+        )
+        .await
+        .unwrap()
+        .json()
+        .unwrap();
+
+    assert_eq!(token_account_id, token_account.id().to_string());
+
+    // Deploy new version
+    let factory = factory
+        .as_account()
+        .deploy(&(std::fs::read(FACTORY_WASM_PATH).unwrap()))
+        .await
+        .unwrap()
+        .result;
+
+    let token_account_id: String = factory
+        .view(
+            "get_bridge_token_account_id",
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes(),
+        )
+        .await
+        .unwrap()
+        .json()
+        .unwrap();
+
+    assert_eq!(token_account_id, token_account.id().to_string());
+
+    // Verify the factory version
+    let factory_version: String = factory
+        .view("version", "".as_bytes().to_vec())
+        .await
+        .unwrap()
+        .json()
+        .unwrap();
+    assert_eq!(factory_version, "0.2.0");
+
+    // Set alice as super admin
+    let result = factory
+        .call("acl_init_super_admin")
+        .args(
+            json!({
+               "account_id": alice.id().to_string()
+            })
+            .to_string()
+            .into_bytes(),
+        )
+        .max_gas()
+        .transact()
+        .await
+        .unwrap();
+
+    let result: bool = result.json().unwrap();
+    assert!(result);
+
+    // Grant alice the `UpgradableManager` role
+    let result = alice
+        .call(factory.id(), "acl_grant_role")
+        .args(
+            json!({
+               "role": "UpgradableManager",
+               "account_id": alice.id().to_string()
+            })
+            .to_string()
+            .into_bytes(),
+        )
+        .max_gas()
+        .transact()
+        .await
+        .unwrap();
+    let result: bool = result.json().unwrap();
+    assert!(result);
+
+    // Upgrade the bridge token directly because self-upgrade is not supported in version `0.1.6`
+    let token_contract = token_account
+        .deploy(&std::fs::read(BRIDGE_TOKEN_WASM_PATH).unwrap())
+        .await
+        .unwrap()
+        .result;
+
+    // Verify the bridge token version
+    let token_version: String = worker
+        .view(token_account.id(), "version", vec![])
+        .await
+        .unwrap()
+        .json()
+        .unwrap();
+    assert_eq!(token_version, "0.2.0");
+
+    // Upgrade the bridge token over factory (redeploy the same version)
+    let result = alice
+        .call(factory.id(), "upgrade_bridge_token")
+        .args(
+            json!({"address": DAI_ADDRESS.to_string()})
+                .to_string()
+                .into_bytes(),
+        )
+        .max_gas()
+        .transact()
+        .await
+        .unwrap();
+    assert!(result.is_success());
+
+    // Verify the bridge token version
+    let token_version: String = worker
+        .view(token_account.id(), "version", vec![])
+        .await
+        .unwrap()
+        .json()
+        .unwrap();
+    assert_eq!(token_version, "0.2.0");
+
+    // Grant alice the `PauseManager` role
+    let result = alice
+        .call(factory.id(), "acl_grant_role")
+        .args(
+            json!({
+               "role": "PauseManager",
+               "account_id": alice.id().to_string()
+            })
+            .to_string()
+            .into_bytes(),
+        )
+        .max_gas()
+        .transact()
+        .await
+        .unwrap();
+    let result: bool = result.json().unwrap();
+    assert!(result);
+
+    // Pause bridge token withdraw
+    let result = alice
+        .call(factory.id(), "set_paused_withdraw")
+        .args(
+            json!({"address": DAI_ADDRESS.to_string(), "paused": true})
+                .to_string()
+                .into_bytes(),
+        )
+        .max_gas()
+        .transact()
+        .await
+        .unwrap();
+    println!("{:?}", result);
+    assert!(result.is_success());
+
+    // Verify bridge token is paused
+    let is_paused: bool = alice
+        .call(token_account.id(), "is_paused")
+        .max_gas()
+        .transact()
+        .await
+        .unwrap()
+        .json()
+        .unwrap();
+    assert!(is_paused);
 }
